project/requirements.txt
Django
djangorestframework
python-dotenv
openai
djongo